bet9ja_v2/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-ai-textanalytics
azure-core
azure-storage-blob
azure-eventhub
